Move Onboarding to new page + wip on Chiev claim - Feedback suggested moving the onboarding page so, here it is - Also moved the effects toggle & connect buttons into the header
